+++ TestSuite/script.py
@@ -1,6 +1,5 @@
 from pathlib import Path
 
-# Do not let GFRUEND change this
 import yaml
 from TestSuite.integration import Integration
 from TestSuite.test_tools import suite_join_path
@@ -20,7 +19,7 @@ class Script(Integration):
         description = open(suite_join_path(default_script_dir, 'sample_script_description.md'))
         self.build(
             code=str(code.read()),
-            yml=yaml.load(yml),
+            yml=yaml.load(yml, Loader=yaml.FullLoader),
             image=image.read(),
             changelog=str(changelog.read()),
             description=str(description.read())