+++ demisto_sdk/commands/common/tests/base_validator_test.py
@@ -59,8 +59,8 @@ def test_check_deprecated_where_ignored_list_exists(repo):
     """
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
-    integration.write_yml({'deprecated': True})
-    files_path = integration.yml_path
+    integration.yml.write_dict({'deprecated': True})
+    files_path = integration.yml.path
     with ChangeCWD(repo.path):
         base_validator = BaseValidator(ignored_errors={'integration.yml': ['BA101']})
         base_validator.check_deprecated(files_path)
@@ -81,8 +81,8 @@ def test_check_deprecated_where_ignored_list_does_not_exist(repo):
     """
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
-    integration.write_yml({'deprecated': True})
-    files_path = integration.yml_path
+    integration.yml.write_dict({'deprecated': True})
+    files_path = integration.yml.path
     with ChangeCWD(repo.path):
         base_validator = BaseValidator(ignored_errors={})
         base_validator.check_deprecated(files_path)
@@ -103,8 +103,8 @@ def test_check_deprecated_non_deprecated_integration_no_ignored_errors(repo):
     """
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
-    integration.write_yml({'deprecated': False})
-    files_path = integration.yml_path
+    integration.yml.write_dict({'deprecated': False})
+    files_path = integration.yml.path
     with ChangeCWD(repo.path):
         base_validator = BaseValidator(ignored_errors={})
         base_validator.check_deprecated(files_path)
@@ -125,8 +125,8 @@ def test_check_deprecated_non_deprecated_integration_with_ignored_errors(repo):
     """
     pack = repo.create_pack('pack')
     integration = pack.create_integration('integration')
-    integration.write_yml({'deprecated': False})
-    files_path = integration.yml_path
+    integration.yml.write_dict({'deprecated': False})
+    files_path = integration.yml.path
     with ChangeCWD(repo.path):
         base_validator = BaseValidator(ignored_errors={'integration.yml': ["BA101"]})
         base_validator.check_deprecated(files_path)
@@ -146,8 +146,8 @@ def test_check_deprecated_playbook(repo):
     """
     pack = repo.create_pack('pack')
     playbook = pack.create_integration('playbook-somePlaybook')
-    playbook.write_yml({'hidden': True})
-    files_path = playbook.yml_path
+    playbook.yml.write_dict({'hidden': True})
+    files_path = playbook.yml.path
     with ChangeCWD(repo.path):
         base_validator = BaseValidator(ignored_errors={})
         base_validator.check_deprecated(files_path)