+++ demisto_sdk/commands/find_dependencies/tests/find_dependencies_test.py
@@ -5,6 +5,7 @@ import pytest
 from demisto_sdk.commands.common.git_tools import git_path
 from demisto_sdk.commands.find_dependencies.find_dependencies import \
     PackDependencies
+from TestSuite.utils import IsEqualFunctions
 
 
 @pytest.fixture(scope="module")
@@ -33,7 +34,7 @@ class TestIdSetFilters:
 
     def test_search_for_specific_pack_script_item(self, id_set):
         pack_id = "PrismaCloudCompute"
-        found_filtered_result = PackDependencies._search_for_pack_items(pack_id, id_set['scripts'])
+
         expected_result = [
             {
                 "PrismaCloudComputeParseAuditAlert": {
@@ -65,7 +66,9 @@ class TestIdSetFilters:
             }
         ]
 
-        assert found_filtered_result == expected_result
+        found_filtered_result = PackDependencies._search_for_pack_items(pack_id, id_set['scripts'])
+
+        assert IsEqualFunctions.is_lists_equal(found_filtered_result, expected_result)
 
     @pytest.mark.parametrize("pack_id", ["Claroty", "Code42", "Cymulate"])
     def test_search_for_pack_playbook_item(self, pack_id, id_set):
@@ -75,7 +78,7 @@ class TestIdSetFilters:
 
     def test_search_for_specific_pack_playbook_item(self, id_set):
         pack_id = "Expanse"
-        found_filtered_result = PackDependencies._search_for_pack_items(pack_id, id_set['playbooks'])
+
         expected_result = [
             {
                 "ExpanseParseRawIncident": {
@@ -85,12 +88,17 @@ class TestIdSetFilters:
                     "implementing_scripts": [
                         "ExpanseParseRawIncident"
                     ],
+                    "tests": [
+                        "No tests (auto formatted)"
+                    ],
                     "pack": "Expanse"
                 }
             }
         ]
 
-        assert found_filtered_result == expected_result
+        found_filtered_result = PackDependencies._search_for_pack_items(pack_id, id_set['playbooks'])
+
+        assert IsEqualFunctions.is_lists_equal(found_filtered_result, expected_result)
 
 
 class TestDependsOnScriptAndIntegration:
@@ -112,9 +120,10 @@ class TestDependsOnScriptAndIntegration:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
 
-        assert found_result == expected_result
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     @pytest.mark.parametrize("dependency_integration_command,expected_result",
                              [("sslbl-get-indicators", {("Feedsslabusech", True)}),
@@ -134,11 +143,14 @@ class TestDependsOnScriptAndIntegration:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
 
-        assert found_result == expected_result
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     def test_collect_scripts_depends_on_two_scripts(self, id_set):
+        expected_result = {('HelloWorld', True), ('PrismaCloudCompute', True)}
+
         test_input = [
             {
                 "DummyScript": {
@@ -152,11 +164,14 @@ class TestDependsOnScriptAndIntegration:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
 
-        assert found_result == {('HelloWorld', True), ('PrismaCloudCompute', True)}
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     def test_collect_scripts_depends_on_two_integrations(self, id_set):
+        expected_result = {('Active_Directory_Query', True), ('Feedsslabusech', True)}
+
         test_input = [
             {
                 "DummyScript": {
@@ -170,11 +185,14 @@ class TestDependsOnScriptAndIntegration:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
 
-        assert found_result == {('Active_Directory_Query', True), ('Feedsslabusech', True)}
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     def test_collect_scripts_depends_on_with_two_inputs(self, id_set):
+        expected_result = {('Active_Directory_Query', True), ('Feedsslabusech', True)}
+
         test_input = [
             {
                 "DummyScript1": {
@@ -200,7 +218,7 @@ class TestDependsOnScriptAndIntegration:
 
         found_result = PackDependencies._collect_scripts_dependencies(pack_scripts=test_input, id_set=id_set)
 
-        assert found_result == {('Active_Directory_Query', True), ('Feedsslabusech', True)}
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     @pytest.mark.parametrize("generic_command", ["ip", "domain", "url"])
     def test_collect_detection_of_optional_dependencies(self, generic_command, id_set):
@@ -252,9 +270,10 @@ class TestDependsOnPlaybook:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
 
-        assert found_result == expected_result
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     @pytest.mark.parametrize("dependency_playbook,expected_result",
                              [("Pentera Run Scan", {("Pcysys", True)}),
@@ -282,9 +301,10 @@ class TestDependsOnPlaybook:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
 
-        assert found_result == expected_result
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     @pytest.mark.parametrize("integration_command,expected_result",
                              [("aws-get-indicators", {("FeedAWS", True)}),
@@ -312,9 +332,10 @@ class TestDependsOnPlaybook:
                 }
             }
         ]
+
         found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input, id_set=id_set)
 
-        assert found_result == expected_result
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
     def test_collect_playbooks_dependencies_on_integrations_with_brand(self, id_set):
         command = "ip"
@@ -375,6 +396,288 @@ class TestDependsOnPlaybook:
 
         for found_result in found_result_set:
             assert not found_result[1]  # validate that mandatory is set to False
+
+    def test_collect_playbooks_dependencies_on_incident_fields(self, id_set):
+        expected_result = {("DigitalGuardian", True), ("EmployeeOffboarding", True)}
+        test_input = [
+            {
+                "Dummy Playbook": {
+                    "name": "Dummy Playbook",
+                    "file_path": "dummy_path",
+                    "fromversion": "dummy_version",
+                    "implementing_scripts": [
+                    ],
+                    "implementing_playbooks": [
+                    ],
+                    "command_to_integration": {
+                    },
+                    "tests": [
+                        "dummy_playbook"
+                    ],
+                    "pack": "dummy_pack",
+                    "incident_fields": [
+                        "digitalguardianusername",
+                        "Google Display Name"
+                    ]
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnLayout:
+    def test_collect_layouts_dependencies(self, id_set):
+        """
+        Given
+            - A layout entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the layout depends on.
+        """
+        expected_result = {("FeedMitreAttack", True), ("PrismaCloudCompute", True), ("CommonTypes", True), ("CrisisManagement", True)}
+
+        test_input = [
+            {
+                "Dummy Layout": {
+                    "typeID": "dummy_layout",
+                    "name": "Dummy Layout",
+                    "pack": "dummy_pack",
+                    "kind": "edit",
+                    "path": "dummy_path",
+                    "incident_and_indicator_types": [
+                        "MITRE ATT&CK",
+                        "Prisma Cloud Compute Cloud Discovery"
+                    ],
+                    "incident_and_indicator_fields": [
+                        "indicator_adminname",
+                        "indicator_jobtitle"
+                    ]
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_layouts_dependencies(pack_layouts=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnIncidentField:
+    def test_collect_incident_field_dependencies(self, id_set):
+        """
+        Given
+            - An incident field entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the incident field depends on.
+        """
+        expected_result = {("Expanse", True), ("IllusiveNetworks", True), ("Carbon_Black_Enterprise_Response", True), ("Phishing", True)}
+
+        test_input = [
+            {
+                "Dummy Incident Field": {
+                    "name": "Dummy Incident Field",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "incident_types": [
+                        "Expanse Appearance",
+                        "Illusive Networks Incident"
+                    ],
+                    "scripts": [
+                        "CBLiveFetchFiles",
+                        "CheckEmailAuthenticity"
+                    ]
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_incidents_fields_dependencies(
+            pack_incidents_fields=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnIndicatorType:
+    def test_collect_indicator_type_dependencies(self, id_set):
+        """
+        Given
+            - An indicator type entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the indicator type depends on.
+        """
+        expected_result = {("Feedsslabusech", True), ("AbuseDB", True), ("ActiveMQ", True), ("CommonScripts", True), ("Carbon_Black_Enterprise_Response", True)}
+
+        test_input = [
+            {
+                "Dummy Indicator Type": {
+                    "name": "Dummy Indicator Type",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "integrations": [
+                        "abuse.ch SSL Blacklist Feed",
+                        "AbuseIPDB",
+                        "ActiveMQ"
+                    ],
+                    "scripts": [
+                        "AssignAnalystToIncident",
+                        "CBAlerts"
+                    ]
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_indicators_types_dependencies(
+            pack_indicators_types=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnIntegrations:
+    def test_collect_integration_dependencies(self, id_set):
+        """
+        Given
+            - An integration entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the integration depends on.
+        """
+        expected_result = {("HelloWorld", True), ("Claroty", True), ("EWS", True), ("CrisisManagement", True), ("CommonTypes", True)}
+
+        test_input = [
+            {
+                "Dummy Integration": {
+                    "name": "Dummy Integration",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "classifiers": "HelloWorld",
+                    "mappers": [
+                        "Claroty-mapper",
+                        "EWS v2-mapper"
+                    ],
+                    "incident_types": "HR Ticket",
+                    "indicator_fields": "CommonTypes",
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_integrations_dependencies(
+            pack_integrations=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnIncidentType:
+    def test_collect_incident_type_dependencies(self, id_set):
+        """
+        Given
+            - An incident type entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the incident type depends on.
+        """
+        expected_result = {("AutoFocus", True), ("Volatility", True)}
+
+        test_input = [
+            {
+                "Dummy Incident Type": {
+                    "name": "Dummy Incident Type",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "playbooks": "Autofocus Query Samples, Sessions and Tags",
+                    "scripts": "AnalyzeMemImage"
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_incidents_types_dependencies(
+            pack_incidents_types=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnClassifiers:
+    def test_collect_classifier_dependencies(self, id_set):
+        """
+        Given
+            - A classifier entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the classifier depends on.
+        """
+        expected_result = {("Claroty", True), ("PAN-OS", True), ("Logzio", True)}
+
+        test_input = [
+            {
+                "Dummy Classifier": {
+                    "name": "Dummy Classifier",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "incident_types": [
+                        "Claroty Integrity Incident",
+                        "FirewallUpgrade",
+                        "Logz.io Alert"
+                    ],
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_classifiers_dependencies(
+            pack_classifiers=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
+
+
+class TestDependsOnMappers:
+    def test_collect_mapper_dependencies(self, id_set):
+        """
+        Given
+            - A mapper entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the mapper depends on.
+        """
+        expected_result = {("AccessInvestigation", True), ("CommonTypes", True), ("PrismaCloud", True), ("BruteForce", True)}
+
+        test_input = [
+            {
+                "Dummy Mapper": {
+                    "name": "Dummy Mapper",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "incident_types": [
+                        "Access",
+                        "Authentication",
+                        "AWS CloudTrail Misconfiguration"
+                    ],
+                    "incident_fields": [
+                        "incident_accountgroups",
+                        "incident_accountid"
+                    ],
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_mappers_dependencies(
+            pack_mappers=test_input, id_set=id_set)
+
+        assert IsEqualFunctions.is_sets_equal(found_result, expected_result)
 
 
 class TestDependencyGraph: