+++ demisto_sdk/commands/validate/file_validator.py
@@ -72,7 +72,9 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS, checked_type,
                                                filter_packagify_changes,
                                                find_type, get_pack_name,
                                                get_remote_file, get_yaml,
+                                               has_remote_configured,
                                                is_file_path_in_pack,
+                                               is_origin_content_repo,
                                                print_color, print_error,
                                                print_warning, run_command,
                                                should_file_skip_validation)
@@ -271,14 +273,38 @@ class FilesValidator:
             print_ignored_files=self.print_ignored_files)
 
         if not self.is_circle:
-            files_string = run_command('git diff --name-status --no-merges HEAD')
-            nc_modified_files, nc_added_files, nc_deleted_files, nc_old_format_files = self.get_modified_files(
-                files_string, print_ignored_files=self.print_ignored_files)
+            remote_configured = has_remote_configured()
+            is_origin_demisto = is_origin_content_repo()
+            if remote_configured and not is_origin_demisto:
+                files_string = run_command('git diff --name-status --no-merges upstream/master...HEAD')
+                nc_modified_files, nc_added_files, nc_deleted_files, nc_old_format_files = \
+                    self.get_modified_files(files_string, print_ignored_files=self.print_ignored_files)
 
-            all_changed_files_string = run_command('git diff --name-status {}'.format(tag))
-            modified_files_from_tag, added_files_from_tag, _, _ = \
-                self.get_modified_files(all_changed_files_string,
-                                        print_ignored_files=self.print_ignored_files)
+                all_changed_files_string = run_command(
+                    'git diff --name-status upstream/master...HEAD')
+                modified_files_from_tag, added_files_from_tag, _, _ = \
+                    self.get_modified_files(all_changed_files_string,
+                                            print_ignored_files=self.print_ignored_files)
+            else:
+                if (not is_origin_demisto and not remote_configured) and self.silence_init_prints:
+                    print_warning(
+                        "Warning: The changes may fail validation once submitted via a "
+                        "PR. To validate your changes, please make sure you have a git remote setup"
+                        " and pointing to github.com/demisto/content.\nYou can do this by running "
+                        "the following commands:\n\ngit remote add upstream https://github.com/"
+                        "demisto/content.git\ngit fetch upstream\n\nMore info about configuring "
+                        "a remote for a fork is available here: https://help.github.com/en/"
+                        "github/collaborating-with-issues-and-pull-requests/configuring-a-"
+                        "remote-for-a-fork"
+                    )
+                files_string = run_command('git diff --name-status --no-merges HEAD')
+                nc_modified_files, nc_added_files, nc_deleted_files, nc_old_format_files = self.get_modified_files(
+                    files_string, print_ignored_files=self.print_ignored_files)
+
+                all_changed_files_string = run_command('git diff --name-status {}'.format(tag))
+                modified_files_from_tag, added_files_from_tag, _, _ = \
+                    self.get_modified_files(all_changed_files_string,
+                                            print_ignored_files=self.print_ignored_files)
 
             if self.file_path:
                 if F'M\t{self.file_path}' in files_string:
@@ -540,7 +566,7 @@ class FilesValidator:
             print('\nValidating {}'.format(file_path))
             self.check_for_spaces_in_file_name(file_path)
 
-            if re.search(TEST_PLAYBOOK_REGEX, file_path, re.IGNORECASE) and not file_type:
+            if re.search(TEST_PLAYBOOK_REGEX, file_path, re.IGNORECASE):
                 continue
 
             elif ('ReleaseNotes' in file_path) and not self.skip_pack_rn_validation: