+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
-* *Unify* command now adds contributor details to the display name and description.
+
+#### 1.1.4
+* Fixed an issue with the default output file name of the **unify** command when using "." as an output path.
+* **Unify** command now adds contributor details to the display name and description.
+* **Format** command now adds *isFetch* and *incidenttype* fields to integration yml.
+* Removed the *feedIncremental* field from the integration schema.
+* **Format** command now adds *feedBypassExclusionList*, *Fetch indicators*, *feedReputation*, *feedReliability*,
+     *feedExpirationPolicy*, *feedExpirationInterval* and *feedFetchInterval* fields to integration yml.
+* Fixed an issue in the playbooks schema.
+* Fixed an issue where generated release notes were out of order.
+* Improved pack dependencies detection.
+* Fixed an issue where test playbooks were mishandled in **validate** command.
 
 #### 1.1.3
 * Added a validation for invalid id fields in indicators types files in **validate** command.